Go setup will now use the setup-go/v4 action
The setup go action has the ability to derive the go version
from the go.mod file so now we only need to define the version
in one place instead of updating the go mod file as well as an action
seperately.
There are other benefits to be had from using the action e.a build in
caching for both the build cache as well as module cache.

Buf has been updated to the latest version