[SPARK-48777][BUILD] Migrate build system to Bazel
### What changes were proposed in this pull request?
Until now the repository was using a hand-written Makefile to compile and build the Spark Connect go client. This comes with some downsides of how to generate for example the protobuf and grpc dependencies and required us to commit the generated code to the repository.

This patch fixes this by migrating the build to Bazel and leveraging the built-in protobuf generation.

### Why are the changes needed?
Build stability

### Does this PR introduce _any_ user-facing change?
NO

### How was this patch tested?
Existing Tests

Closes #23 from grundprinzip/SPARK-48777.

Authored-by: Martin Grund <martin.grund@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>