[SPARK-48754] Support Retries in Go client
### What changes were proposed in this pull request?

Adding support for retriable executions in the Spark Connect Go client. In addition, this is the last big refactoring that is needed to produce a more structured and cleaner code-base in general.

It drives the separation of GRPC client and the actual Spark operations that are executed on top of it. In addition, it homogenized the naming of the different types so that they're, in general, closer to the other Spark Connect client implementations.

### Why are the changes needed?
Production Support

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Added many new tests.

Closes #39 from grundprinzip/retries.

Authored-by: Martin Grund <martin.grund@databricks.com>
Signed-off-by: Martin Grund <martin.grund@databricks.com>