Cleanup row type and Arrow conversion
### What changes were proposed in this pull request?
This patch removes some of the unnecessary indirections of the previous implementation of the Row type and addss support for deserializing LIST types from Arrow as well.

### Why are the changes needed?
Compatibility / Ease of Use

### Does this PR introduce _any_ user-facing change?
Now the `Row` type is closer to the PySpark one avoiding the unnecessary extraction to `row.values()` to access the values by offset or name.

### How was this patch tested?
Added Unit tests

Closes #72 from grundprinzip/cleanup_row_type.

Authored-by: Martin Grund <martin.grund@databricks.com>
Signed-off-by: Martin Grund <martin.grund@databricks.com>