Simplified the code, and added GetWithDefault
Instead of sending over the ConfigRequest, I changed it to send over the Operation and ensure that I create the ConfigRequest in the client.

Furthermore I added verification that the value returned is not nil
Finally I added GetWithDefault which takes a map[string]string with the key and the default value if the key is not set.